// scn-redactor/src/app/shared/sc-edge-idtf.enum.ts
export enum ScEdgeIdtf {
    EdgeUCommonConst = "EdgeUCommonConst",
    EdgeDCommonConst = "EdgeDCommonConst",
    EdgeUCommonVar = "EdgeUCommonVar",
    EdgeDCommonVar = "EdgeDCommonVar",
    EdgeAccessConstPosPerm = "EdgeAccessConstPosPerm",
    EdgeAccessConstNegPerm = "EdgeAccessConstNegPerm",
    EdgeAccessConstFuzPerm = "EdgeAccessConstFuzPerm",
    EdgeAccessConstPosTemp = "EdgeAccessConstPosTemp",
    EdgeAccessConstNegTemp = "EdgeAccessConstNegTemp",
    EdgeAccessConstFuzTemp = "EdgeAccessConstFuzTemp",
    EdgeAccessVarPosPerm = "EdgeAccessVarPosPerm",
    EdgeAccessVarNegPerm = "EdgeAccessVarNegPerm",
    EdgeAccessVarFuzPerm = "EdgeAccessVarFuzPerm",
    EdgeAccessVarPosTemp = "EdgeAccessVarPosTemp",
    EdgeAccessVarNegTemp = "EdgeAccessVarNegTemp",
    EdgeAccessVarFuzTemp = "EdgeAccessVarFuzTemp",
}